corrected media queries and client side navigation
